feat: :sparkles: http responses

+++ proyectospringboot/src/main/java/com/campuslands/proyectospringboot/Barrio/infrastructure/in/web/BarrioController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -18,8 +19,6 @@ import com.campuslands.proyectospringboot.Barrio.application.BarrioService;
 import com.campuslands.proyectospringboot.Barrio.domain.Barrio;
 
 import jakarta.validation.Valid;
-
-import org.springframework.web.bind.annotation.PutMapping;
 
 @RestController
 @RequestMapping("/barrio")
@@ -49,13 +48,23 @@ public class BarrioController {
     }
 
     @PutMapping("/{id}")
-    public void update(@PathVariable Long id, @Valid @RequestBody Barrio barrio) {
+    public ResponseEntity<String> updateBarrio(@PathVariable Long id, @Valid @RequestBody Barrio barrio) {
+        Optional<Barrio> foundBarrio = barrioService.findById(id);
+        if (!foundBarrio.isPresent()) {
+            return new ResponseEntity<>("Barrio no encontrado", HttpStatus.NOT_FOUND);
+        }
         barrio.setId(id);
         barrioService.create(barrio);
+        return new ResponseEntity<>("Barrio ha sido guardado correctamente", HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteBarrio(@PathVariable Long id) {
+        Optional<Barrio> foundBarrio = barrioService.findById(id);
+        if (!foundBarrio.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         barrioService.deleteById(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
