+++ proyectospringboot/src/main/java/com/campuslands/proyectospringboot/Dimensiones/infrastructure/adapters/in/DimensionesController.java
@@ -1,7 +1,10 @@
 package com.campuslands.proyectospringboot.Dimensiones.infrastructure.adapters.in;
 
 import java.util.List;
+import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.campuslands.proyectospringboot.Dimensiones.app.services.DimensionesService;
 import com.campuslands.proyectospringboot.Dimensiones.domain.entities.Dimensiones;
 
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 
 @RestController
@@ -28,8 +32,12 @@ public class DimensionesController {
     }
 
     @GetMapping("/{id}")
-    public Dimensiones getDimensionesById(@PathVariable Long id) {
-        return dimensionesService.findById(id);
+    public ResponseEntity<Dimensiones> listById(@PathVariable Long id) {
+        Optional<Dimensiones> fDimensiones = dimensionesService.findById(id);
+        if (!fDimensiones.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(fDimensiones.get(), HttpStatus.OK);
     }
 
     @PostMapping
@@ -38,13 +46,23 @@ public class DimensionesController {
     }
 
     @PutMapping("/{id}")
-    public Dimensiones updateDimensiones(@PathVariable Long id, @RequestBody Dimensiones dimensiones) {
-        dimensiones.setId(id);
-        return dimensionesService.save(dimensiones);
+    public ResponseEntity<String> updateDimensiones(@PathVariable Long id, @Valid @RequestBody Dimensiones Dimensiones) {
+        Optional<Dimensiones> foundDimensiones = dimensionesService.findById(id);
+        if (!foundDimensiones.isPresent()) {
+            return new ResponseEntity<>("Dimensiones no encontrado", HttpStatus.NOT_FOUND);
+        }
+        Dimensiones.setId(id);
+        dimensionesService.save(Dimensiones);
+        return new ResponseEntity<>("Dimensiones ha sido guardado correctamente", HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteDimensiones(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteDimensiones(@PathVariable Long id) {
+        Optional<Dimensiones> foundDimensiones = dimensionesService.findById(id);
+        if (!foundDimensiones.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         dimensionesService.deleteById(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
