+++ proyectospringboot/src/main/java/com/campuslands/proyectospringboot/Direccion/infrastructure/adapters/in/DireccionController.java
@@ -1,7 +1,10 @@
 package com.campuslands.proyectospringboot.Direccion.infrastructure.adapters.in;
 
 import java.util.List;
+import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,8 +33,12 @@ public class DireccionController {
     }
 
     @GetMapping("/{id}")
-    public Direccion getDireccionById(@PathVariable Long id) {
-        return direccionService.findById(id);
+    public ResponseEntity<Direccion> listById(@PathVariable Long id) {
+        Optional<Direccion> fDireccion = direccionService.findById(id);
+        if (!fDireccion.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(fDireccion.get(), HttpStatus.OK);
     }
 
     @PostMapping
@@ -40,13 +47,23 @@ public class DireccionController {
     }
 
     @PutMapping("/{id}")
-    public Direccion updateDireccion(@PathVariable Long id, @Valid @RequestBody Direccion direccion) {
-        direccion.setId(id);
-        return direccionService.save(direccion);
+    public ResponseEntity<String> updateDireccion(@PathVariable Long id, @Valid @RequestBody Direccion Direccion) {
+        Optional<Direccion> foundDireccion = direccionService.findById(id);
+        if (!foundDireccion.isPresent()) {
+            return new ResponseEntity<>("Direccion no encontrado", HttpStatus.NOT_FOUND);
+        }
+        Direccion.setId(id);
+        direccionService.save(Direccion);
+        return new ResponseEntity<>("Direccion ha sido guardado correctamente", HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteDireccion(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteDireccion(@PathVariable Long id) {
+        Optional<Direccion> foundDireccion = direccionService.findById(id);
+        if (!foundDireccion.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         direccionService.deleteById(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
